FilterLib: addressing compiler warnings, Xcode warnings, and removing unneeded casting of values.

+++ DSP/extensions/FilterLib/includes/TTFilter.h
@@ -56,7 +56,7 @@ public:
 			if (err == kTTErrInvalidAttribute)
 				err = mActualFilterObject->setAttributeValue(TT("resonance"), mQ);
 			mActualFilterObject->setAttributeValue(TT("bypass"), this->attrBypass);
-			mActualFilterObject->setAttributeValue(kTTSym_sampleRate, (unsigned int)sr);
+			mActualFilterObject->setAttributeValue(kTTSym_sampleRate, sr);
 		}
 		return err;
 	}
@@ -94,7 +94,7 @@ public:
 	
 	TTErr updateSampleRate(const TTValue& oldSampleRate, TTValue&)
 	{
-		return mActualFilterObject->setAttributeValue(kTTSym_sampleRate, (unsigned int)sr);
+		return mActualFilterObject->setAttributeValue(kTTSym_sampleRate, sr);
 	}
